Initial commit on Joystick widget.
For a number of games I have planed, I need an onscreen joystick.
So far I have actually managed to get most of the behavior I'm
looking for programmed in. The last piece of behavior that is
needed is what happens when a touch starts inside the joystick area
and moves out. I tried implementing it tonight, but ran into some
bugs. It gets a little funky. Not sure where my math went wrong.

There will also need to be some work done to generalize it a little
more. Right now it basically has to be 200x200. I've put in some
calculations that attempt to deal with arbitrary sizes, but getting
a function joystick has been my first priority of course.