Fixed the maths!
I fixed the maths. The main the was that I was assigning it to
the pad's x and y verses pad's center. I also had to deal with a
flipped sign. Basically subtracting the widget's center from the
touch rather than the other way around (despite that being how I
left that).