Finished it! Mostly at least.
I've got a little bit of polishing left to do on the code. Mostly
it will be removing commented code and removing debug print
statements, oh and adding comments for maintainability. But beyond
that it's basically done. Now I need to figure out how to share it
via the Kivy Garden.