Preped for Kivy Gardening
I believe I've got this already now for Kivy Garden. It was a bit
of a thing to try and get the pep8 checker working. I basically
just copied a number of files from the kivy designer over to my
dist-packages directory. I actually had to make the kivy/tools/
pep8checker directory in order to put the files there. It was a bit
of a mess.

But I can finally commit, as you can see. I did also make a few
changes worth noting. I fixed the issue where on touch move would
be picked up by the joystick even if it wasn't the joystick's
touch that moved.